% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_fit_info.R
\name{attach_fit_info}
\alias{attach_fit_info}
\title{Attach fit info and tables to a fit object, e.g. from model fit or
Pharmpy grid search final results}
\usage{
attach_fit_info(
  fit,
  model,
  fit_folder,
  output_file = file.path(fit_folder, "model.lst"),
  verbose = TRUE
)
}
\description{
Attach fit info and tables to a fit object, e.g. from model fit or
Pharmpy grid search final results
}
