% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{calc_pk_variables}
\alias{calc_pk_variables}
\title{Calculate some basic PK variables from simulated or observed data}
\usage{
calc_pk_variables(data, regimen = NULL, dictionary = NULL)
}
\arguments{
\item{data}{data.frame in NONMEM format}

\item{regimen}{if specified, will replace the regimens for each subject with
a custom regimen, specified using arguments \code{dose}, \code{interval}, \code{n}, and
\code{route} (and \code{t_inf} / \code{rate} for infusions).
E.g. \code{regimen = list(dose = 500, interval = 12, n = 5, route = "oral")}.}
}
\value{
data.frame
}
\description{
Calculate some basic PK variables from simulated or observed data
}
