% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_psn.R
\name{call_psn}
\alias{call_psn}
\title{Call PsN}
\usage{
call_psn(
  model_file,
  output_file,
  path,
  tool = c("execute", "vpc", "bootstrap", "sir", "proseval"),
  options = NULL,
  console = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{model_file}{model file, e.g. "run.mod"}

\item{output_file}{output file, e.g. "run.lst"}

\item{path}{run folder path, e.g. "run1"}

\item{console}{show output from nmfe in console? Default \code{FALSE}}

\item{verbose}{verbose output?}
}
\description{
Call PsN
}
