% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_modelfit_data.R
\name{clean_modelfit_data}
\alias{clean_modelfit_data}
\title{Clean / check the dataset before passing to model fitting tool}
\usage{
clean_modelfit_data(model, data = NULL)
}
\description{
Clean / check the dataset before passing to model fitting tool
}
