% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_nlme_fit.R
\name{combine_info_columns}
\alias{combine_info_columns}
\title{Combine columns with run info into a data.frame
and make sure that rows match (e.g. parameters)}
\usage{
combine_info_columns(fit_info, table = "info_tab", label = "Detail")
}
\description{
data.frames in list should have the same column names but can have different
row names (e.g. parameter names).
}
