% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_nlme_fit.R
\name{compare_nlme_fit}
\alias{compare_nlme_fit}
\title{Compare fit of two or more NLME fits}
\usage{
compare_nlme_fit(..., return_object = FALSE)
}
\arguments{
\item{...}{fit objects}

\item{return_object}{logical, if TRUE, return a list of the combined info and parameter tables}
}
\description{
Compare fit of two or more NLME fits
}
