% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariate_search_space.R
\name{create_covariate_search_space}
\alias{create_covariate_search_space}
\title{Create covariate search space definition for pharmpy cov_search}
\usage{
create_covariate_search_space(
  parameters,
  covariates,
  operation = c("*", "+", "EXP", "POW"),
  explore = TRUE,
  struct_parameters = NULL,
  struct_covariates = NULL,
  struct_operation = c("*", "+", "EXP", "POW")
)
}
\description{
Create covariate search space definition for pharmpy cov_search
}
