% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{create_dosing_records.data.frame}
\alias{create_dosing_records.data.frame}
\title{Create dosing records, given a specified regimen as a data frame with
potentially multiple regimens and varying dosing times / doses}
\usage{
\method{create_dosing_records}{data.frame}(regimen, data, n_subjects, dictionary)
}
\description{
Create dosing records, given a specified regimen as a data frame with
potentially multiple regimens and varying dosing times / doses
}
