% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{create_dosing_records.list}
\alias{create_dosing_records.list}
\title{Create dosing records, given a specified regimen as a simple list}
\usage{
\method{create_dosing_records}{list}(regimen, data, n_subjects, dictionary)
}
\description{
Create dosing records, given a specified regimen as a simple list
}
