% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{create_dosing_records}
\alias{create_dosing_records}
\title{Create dosing records, given a specified regimen}
\usage{
create_dosing_records(regimen, data, n_subjects, dictionary)
}
\description{
Create dosing records, given a specified regimen
}
