% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_luna_cache.R
\name{create_luna_cache}
\alias{create_luna_cache}
\title{Create a cache environment to store settings for project and system}
\usage{
create_luna_cache(folder, name = NULL, verbose = FALSE)
}
\arguments{
\item{folder}{path to project}

\item{name}{optional, name of project (matching to yaml file)}
}
\description{
Create a cache environment to store settings for project and system
}
