% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model.old}
\alias{create_model.old}
\title{Create a model (old version)}
\usage{
create_model.old(
  n_cmt = 1,
  route = NULL,
  absorption = c("linear", "linear_lag", "transit"),
  software = c("nlmixr2", "nlmixr"),
  data
)
}
\arguments{
\item{n_cmt}{number of compartments for the population PK model}

\item{route}{administration route, either \code{"oral"} or \code{"iv"} or NULL. If NULL
(default) will read from data (\code{ROUTE} column) and used route specified for
first dose in dataset.}

\item{absorption}{one of \code{"linear"}, \code{"linear_lag"} (linear with lagtime), or
\code{"transit"} (transit compartments using the Stirling approximation as
described in Savic et al. JPKPD 2007). Argument is only required when
\code{route} is \code{"oral"}.}

\item{software}{Currently only \code{nlmixr2} supported.}
}
\value{
an R function object for nlmixr2, or a NONMEM model as a
character vector (TODO)
}
\description{
Create a model (old version)
}
