% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nmfe_location_for_run.R
\name{detect_nmfe_version}
\alias{detect_nmfe_version}
\title{get nmfe file name from a NONMEM installation folder}
\usage{
detect_nmfe_version(nm_path)
}
\description{
get nmfe file name from a NONMEM installation folder
}
