\name{download_ers}

\alias{download_ers}

\title{download from ERS}

\description{
Find and download ee data 
}

\usage{
download_ers(scenes, path, username, password, overwrite=FALSE, ...)
}

\arguments{
  \item{scenes}{character}
  \item{path}{character. Path name indicating where to store the data}
  \item{username}{character. User name (see Details)}
  \item{password}{character. Password (see Details)}
  \item{overwrite}{logical. If \code{FALSE}, existing files are overwritten}
  \item{...}{additional arguments. None implemented}
}


\value{
Character vector of file names pointing to the downloaded files
}

\details{
to do
}

\seealso{\code{\link{getLandsat}}} 

\examples{  
# to do
}

