% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{find_file_with_fallback}
\alias{find_file_with_fallback}
\title{Find a file(s) from a model run with a potential fallback}
\usage{
find_file_with_fallback(
  folder,
  filename,
  fallback,
  verbose = TRUE,
  abort = TRUE
)
}
\arguments{
\item{folder}{main model folder}

\item{filename}{by default will look for NONMEM nmfe results file, but can also
look e.g. for a table. In that case specify \code{filename}.}

\item{fallback}{fallback filename}
}
\description{
Find a file(s) from a model run with a potential fallback
}
