% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs_as_table.R
\name{get_all_results}
\alias{get_all_results}
\title{helper function to extract all fit results based on
vector of model files}
\usage{
get_all_results(model_file)
}
\description{
helper function to extract all fit results based on
vector of model files
}
