% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cov_matrix.R
\name{get_cov_matrix}
\alias{get_cov_matrix}
\title{Create a covariance block matrix}
\usage{
get_cov_matrix(params, keep_all = FALSE, triangle = FALSE, nonmem = TRUE)
}
\arguments{
\item{params}{parameters, a vector of standard deviations and correlations,
e.g. \code{list("CL" = 0.1, "V" = 0.2, "KA" = 0.3, "CL~V" = 0.3)}.}

\item{keep_all}{Should all parameters be kept in the covariance matrix,
even if they do not have a correlation with other parameters?}

\item{triangle}{return the lower triangle as a vector instead of a
matrix object?}
}
\description{
Create a covariance block matrix
}
\examples{
\dontrun{
make_cov_matrix(list("CL" = 0.1, "V" = 0.2, "KA" = 0.3, "CL~V" = 0.3))
}

}
