% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_flag_from_config}
\alias{get_flag_from_config}
\title{Get a binary value from config, based on path}
\usage{
get_flag_from_config(
  flag = c("tools", "modelfit", "force"),
  default = FALSE,
  config = NULL
)
}
\arguments{
\item{flag}{character vector indicating path to flag}

\item{default}{default value for flag}

\item{config}{list of config settings. If \code{NULL} will get it using
\code{get_luna_config()}}
}
\description{
Get a binary value from config, based on path
}
