% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_pharmpy_rundolder.R
\name{get_last_pharmpy_runfolder}
\alias{get_last_pharmpy_runfolder}
\title{Find last pharmpy run folder}
\usage{
get_last_pharmpy_runfolder(id, folder = NULL, tool)
}
\arguments{
\item{id}{model id. Optional. If not specified, will generate random modelfit
id. The \code{id} will be used to create the run folder.}
}
\description{
Find last pharmpy run folder
}
