% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_luna_config.R
\name{get_luna_config}
\alias{get_luna_config}
\title{Get luna config, either from global config file,
or from project config}
\usage{
get_luna_config(verbose = TRUE)
}
\arguments{
\item{verbose}{verbose output?}
}
\description{
Get luna config, either from global config file,
or from project config
}
