% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_random_id}
\alias{get_random_id}
\title{Get a random sequence of letters and numbers of size \code{n}}
\usage{
get_random_id(prefix = NULL, n = 6)
}
\arguments{
\item{prefix}{optional prefix}
}
\value{
character
}
\description{
Get a random sequence of letters and numbers of size \code{n}
}
