% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status.R
\name{get_status}
\alias{get_status}
\title{Get the status of a run}
\usage{
get_status(id, folder = ".")
}
\arguments{
\item{id}{id of run to get status for}

\item{folder}{project folder to look in}
}
\description{
Get the status of a run
}
