% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_for_plots.R
\name{get_table_for_plots}
\alias{get_table_for_plots}
\title{Get table for basic GOF plots based on run id}
\usage{
get_table_for_plots(
  id,
  folder = NULL,
  residual = c("CWRES", "NPDE"),
  ltbs = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{folder}{path to folder containing the model file. Default is current directory.}

\item{residual}{either "CWRES" or "NPDE"}

\item{ltbs}{log-transform-both-sides error model? If \code{TRUE}, will
exponentiate DV, PRED, and IPRED.}
}
\description{
Get table for basic GOF plots based on run id
}
