% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_time_last_updated_file}
\alias{get_time_last_updated_file}
\title{Get date/time stamp for last update to file}
\usage{
get_time_last_updated_file(file)
}
\value{
datetime stamp in ISO8601 format. Returns empty string if file
does not exist
}
\description{
Get date/time stamp for last update to file
}
