% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_time_last_updated_folder}
\alias{get_time_last_updated_folder}
\title{Get a timestamp for when the last update was made to any file in a folder
Will look only 1 level deep.}
\usage{
get_time_last_updated_folder(folder)
}
\value{
datetime stamp in ISO8601 format. Returns empty string if no
files in folder
}
\description{
Get a timestamp for when the last update was made to any file in a folder
Will look only 1 level deep.
}
