% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tools_info.R
\name{get_tools_info}
\alias{get_tools_info}
\title{For a given run, get info on what folders are available for selected
Pharmpy or PsN tools}
\usage{
get_tools_info(id, folder, tools = c())
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{folder}{path to folder containing the model file. Default is current directory.}

\item{tools}{vector of tools for which to find info}
}
\description{
For a given run, get info on what folders are available for selected
Pharmpy or PsN tools
}
