% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{handle_system_errors}
\alias{handle_system_errors}
\title{Handle errors from running console commands}
\usage{
handle_system_errors(response, command)
}
\arguments{
\item{response}{response from system2}

\item{command}{command used}
}
\description{
Handle errors from running console commands
}
