% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{insert_entry}
\alias{insert_entry}
\title{The reverse of pluck_entry, insert_entry() inserts
an entry into an unnamded outer list, based on an element
in the inner list.}
\usage{
insert_entry(x, id, entry, el = "id")
}
\description{
The reverse of pluck_entry, insert_entry() inserts
an entry into an unnamded outer list, based on an element
in the inner list.
}
