% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoc.R
\name{is.outlier.inc}
\alias{is.outlier.inc}
\title{Check outliers with masking}
\usage{
is.outlier.inc(x, inc = inc, t = 3)
}
\arguments{
\item{x}{numeric vector}

\item{inc}{boolean vector with same length as \code{x}, used for masking.
\code{FALSE} will exclude the element to \code{NA} in the corresponding position
in \code{x} from calculation.}

\item{t}{scale of deviation, default to 3, as in \code{\link{outliers}}}
}
\value{
boolean vector with same length as \code{x}, \code{TRUE} means the corresponding
element in \code{x} is an outlier and is not \code{NA}, or it is excluded by \code{inc}
}
\description{
Check outliers with masking
}
