% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoc.R
\name{is.outlier}
\alias{is.outlier}
\title{Check outliers}
\usage{
is.outlier(x, m = mean(x, na.rm = T), sdev = sd(x, na.rm = T), t = 3)
}
\arguments{
\item{x}{numeric vector}

\item{m}{target center of acceptable interval after outlier removal,
default to \code{x}'s mean}

\item{sdev}{target variation of acceptable interval, default to \code{x}'s
standard deviation}

\item{t}{scale of deviation, default to 3}
}
\value{
boolean vector with same length as \code{x}, \code{TRUE} means the corresponding
element in \code{x} is an outlier and is not \code{NA}
}
\description{
Check outliers
}
