% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_misc.R
\name{is_luna_cache_available}
\alias{is_luna_cache_available}
\title{Check if luna cache is available / loaded}
\usage{
is_luna_cache_available(abort = FALSE)
}
\description{
Check if luna cache is available / loaded
}
