% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_maxeval_zero}
\alias{is_maxeval_zero}
\title{Does the last estimation method in a model have maxeval=0?}
\usage{
is_maxeval_zero(model)
}
\description{
Does the last estimation method in a model have maxeval=0?
}
