% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_run_as_job}
\alias{is_run_as_job}
\title{Checks config if should run as job}
\usage{
is_run_as_job(config, as_job = NULL)
}
\arguments{
\item{config}{config list}

\item{as_job}{boolean, manual override over config}
}
\description{
Checks config if should run as job
}
