% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ladd.annot.file}
\alias{ladd.annot.file}
\title{Attaches an new annotation from an file}
\usage{
ladd.annot.file(a)
}
\arguments{
\item{a}{a required annotation file name}
}
\value{
no explicit return value
}
\description{
Adds new annotations to the currently attached EDF from a file
}
\note{
This command is similar to specifying an annotation file in a sample-list.
Epoch annotation (.eannot) files must match the attached EDF in terms of the number of current epochs.
}
