% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ladd.annot}
\alias{ladd.annot}
\title{Attaches an new annotation from an R interval list}
\usage{
ladd.annot(annot, intervals)
}
\arguments{
\item{annot}{a required annotation class name for the new annotation}

\item{intervals}{a required list of intervals, in the same format as returned
by \code{\link{lannots}(annot)}}
}
\value{
no explicit return value
}
\description{
Adds new annotations to the currently attached EDF given an interval list
}
\note{
Added annotations will be visible with subsequent \code{\link{lannots}()} commands.
}
