% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lattach}
\alias{lattach}
\title{Attaches an EDF from a sample list}
\usage{
lattach(sl, idx = "")
}
\arguments{
\item{sl}{a sample-list as loaded by \code{\link{lsl}}}

\item{idx}{either an integer number (in which case, it means to attach the
EDF/annotations specified on that row of the sample-list), or a string value
(in which case, it is interpreted as the ID of the individual/EDF to be attached)}
}
\value{
No explicit return value: this command sets the in-memory EDF
representation to reflect this EDF header/file, by calling \code{\link{ledf}}
}
\description{
Loads an EDF and any associated annotation files from a
\href{https://zzz.bwh.harvard.edu/luna/luna/args/#sample-lists}{sample-list}
loaded by \code{\link{lsl}}.
}
\note{
As well as attaching that EDF, this calls \code{\link{lstat}} to display some
basic information about it. If the index or ID is out-of-range/not found, this
command will give an error

Unlike many R functions, \code{lattach()} does not return an object that represents the
data (i.e. the EDF). Rather, \code{lunaR} is designed to operate on one EDF at a time;
attached EDFs can be displayed with the \code{\link{lstat}} function. Attaching a new EDF
effectively detaches any previously attached EDF.
}
\examples{
\dontrun{
> lattach(sl, 2)
nsrr02 : 14 signals, 10 annotations, of 09:57:30 duration

> lattach(sl, "nsrr02")
nsrr02 : 14 signals, 10 annotations, of 09:57:30 duration
}
}
