% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lbands}
\alias{lbands}
\title{Splits a signal into five band-pass filtered signals}
\usage{
lbands(l)
}
\arguments{
\item{l}{a required parameter, which should be a single channel name, e.g.
matching one from \code{\link{lchs}()}}
}
\value{
if the channel \code{l} exists, this command will generate five new EDF
channels that are band-pass filtered versions of the original, for delta,
theta, alpha, sigma and beta bands.
}
\description{
Convenience function to generate five new band-filtered signals given a raw EEG channel
}
\note{
It is simply a wrapper around a series of \href{https://zzz.bwh.harvard.edu/luna/ref/manipulations/#copy}{COPY}
and \href{https://zzz.bwh.harvard.edu/luna/ref/fir-filters/#filter}{FILTER} commands.
You'll need to edit the function definition, or supply a new one, to get
different bands, etc. (Just type \code{lbands} at the R command line to
see the function definition.)
}
