% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lchs}
\alias{lchs}
\title{Returns the EDF channel names}
\usage{
lchs()
}
\value{
a vector of channel names for the attached EDF
}
\description{
Returns a vector of channel names for the attached EDF
}
\examples{
\dontrun{
> lchs()
[1] "SaO2"     "PR"       "EEG(sec)" "ECG"      "EMG"      "EOG(L)"
[7] "EOG(R)"   "EEG"      "AIRFLOW"  "THOR RES" "ABDO RES" "POSITION"
[13] "LIGHT"    "OX STAT"
}
}
