% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lcmd}
\alias{lcmd}
\title{Reads and parses a Luna command script from a file}
\usage{
lcmd(filename)
}
\arguments{
\item{filename}{a required parameter, the name of the file to read}
}
\value{
a vector of Luna commands, parsed from the
\href{https://zzz.bwh.harvard.edu/luna/luna/args/#command-files}{command file}
in a manner that is suitable for \code{\link{leval}()} (and \code{\link{leval.project}()}.
That is, blank lines and comments (starting with \code{\%}) are stripped away,
and multi-line statements are concatenated into a single line.
}
\description{
Reads Luna commands from a file
}
\examples{
k <- leval(lcmd("/path/to/command.txt"))
}
