% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ldata.intervals}
\alias{ldata.intervals}
\title{Extracts signal (and annotation) data from an EDF for one or more intervals}
\usage{
ldata.intervals(i, chs, annots = character(0), w = 0)
}
\arguments{
\item{i}{a required list of intervals (in seconds), e.g. as returned by \code{\link{lannots}(annot)}}

\item{chs}{a required vector of channel names to be returned, i.e. as from \code{\link{lchs}()},
all of which are required to have the same sampling rate is a required vector of channel names}

\item{annots}{an optional vector of annotation names to be returned, i.e. as from \code{\link{lannots}()}}

\item{w}{an optional window (in seconds) the specifies whether the returned
region should also contain a window of up to \code{w} seconds each side of the
specified interval(s). For example, if the intervals represented individual
sleep spindles, then adding \code{w=10} would return windows of approximately 20 seconds,
with the spindle centered in the middle.}
}
\value{
a data frame where each row is one sample point, containing the raw
signal (and annotation) data. See the description of \code{\link{ldata}}'s output for more
information of columns
}
\description{
Returns raw signal and annotation data for requested intervals
}
\note{
This command is effectively the same as \code{\link{ldata}()} except that the regions
of interest are specified as intervals rather than in terms of epoch numbers.
Intervals are defined as half-open intervals [a,b), meaning all sample-points
equal to or greater than a, but less than b.
}
