% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ldb}
\alias{ldb}
\title{Imports data from a lout database as an R list object}
\usage{
ldb(dbfile, ids = character(0))
}
\arguments{
\item{dbfile}{the name of a \href{https://zzz.bwh.harvard.edu/luna/luna/args/#destrat}{lout} database}

\item{ids}{(optional) if \code{ids} is a vector of individual/EDF IDs, then only
these individuals will be extracted from the database}
}
\value{
a named-list R object, in the same format as \code{\link{leval}()}
}
\description{
Imports the contents of a lout database
}
\note{
As noted, the returned value is in the same format as is generated by
\code{\link{leval}()} and can be parsed with \code{\link{lx}()} and \code{\link{lid}()}.
}
\examples{
\dontrun{
> k <- ldb("out.db")
read data on 3 individuals

> k
$HEADERS
$HEADERS$BL
      ID    NR NS REC.DUR TOT.DUR.HMS TOT.DUR.SEC
1 nsrr01 40920 14       1    11:22:00       40920
2 nsrr02 35850 14       1    09:57:30       35850
3 nsrr03 40920 14       1    11:22:00       40920

$HEADERS$CH
       ID       CH  DMAX   DMIN PDIM   PMAX    PMIN  SR
1  nsrr01 ABDO RES   127   -128       -1.00    1.00  10
2  nsrr02 ABDO RES   127   -128       -1.00    1.00  10
3  nsrr03 ABDO RES   127   -128       -1.00    1.00  10
4  nsrr01  AIRFLOW   127   -128       -1.00    1.00  10
5  nsrr02  AIRFLOW   127   -128       -1.00    1.00  10
6  nsrr03  AIRFLOW   127   -128       -1.00    1.00  10
7  nsrr01      ECG   127   -128   mV   1.25   -1.25 250
8  nsrr02      ECG   127   -128   mV   1.25   -1.25 250
9  nsrr03      ECG   127   -128   mV   1.25   -1.25 250
10 nsrr01      EEG   127   -128   uV 125.00 -125.00 125
11 nsrr02      EEG   127   -128   uV 125.00 -125.00 125
12 nsrr03      EEG   127   -128   uV 125.00 -125.00 125
...(continued)...
}

}
