% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ldefault.coh.xy}
\alias{ldefault.coh.xy}
\title{Generating a rescaled coordinate data frame}
\usage{
ldefault.coh.xy(xy)
}
\arguments{
\item{xy}{a data frame with columns \code{X} and \code{Y}}
}
\value{
a data frame with X and Y coordinates rescaled
}
\description{
Rescales a coordinate data frame (e.g. from \code{\link{ldefault.xy}()})
such that it is appropriate for pairwise (connectivity) plots
}
\note{
\itemize{
\item Assumes the X and Y inputs are scaled for a unit circle, i.e. -0.5 to +0.5
\item This function is only intended to be used internally
}
}
