% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ldefault.xy}
\alias{ldefault.xy}
\title{Generating a data frame with default 64-channel EEG positions}
\usage{
ldefault.xy(chs = character(0))
}
\arguments{
\item{chs}{an optional character vector of channel labels}
}
\value{
a data frame of channel, X and Y co-ordiantes (CH, X and Y);  if the
\code{chs} argument is not empty, then only channels matching an entry in \code{chs} are displayed
}
\description{
Generating a data frame with default 64-channel EEG positions
}
\note{
\itemize{
\item Label matching is case-insenstive;
\item Currently, coordinates are only specified for standard 64-channel montages;
\item This function is primarily used internally by other luna plotting functions;
}
}
