% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ldenoise}
\alias{ldenoise}
\title{1D total variation denoiser}
\usage{
ldenoise(x, lambda)
}
\arguments{
\item{x}{a time series vector}

\item{lambda}{Smoothing parameter (0 to infinity)}
}
\value{
a denoised version of \code{x}
}
\description{
Applies a 1D total variation denoiser
}
\note{
This is a wrapper around the method implemented by the \code{TV} command.
See \href{https://zzz.bwh.harvard.edu/luna/ref/power-spectra/#tv}{this page}
for a description of the method and parameters.
}
