% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{le2i}
\alias{le2i}
\title{Converts epochs to intervals}
\usage{
le2i(e, dur = 30, inc = 30)
}
\arguments{
\item{e}{a required parameter, a vector of 1-based epochs}

\item{dur}{(optional) sets the epoch length to be used (default 30 seconds)}

\item{inc}{(optional) set the epoch increment to be used (default 30 seconds)}
}
\value{
an interval list, in the format used by \code{\link{lannots}(annot)} and
\code{\link{ldata.intervals}()}, for example
}
\description{
Returns an interval list corresponding to epochs
}
\examples{
le2i(1:4)
}
