% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ledf}
\alias{ledf}
\title{Attaches an EDF directly}
\usage{
ledf(edf, id = ".", annots = character(0))
}
\arguments{
\item{edf}{a required filename for the to-be-attached EDF}

\item{id}{an optional ID that will be associated with this EDF}

\item{annots}{is an optional vector of one or more annotation filenames
(\code{.xml, .ftr, .annot} or \code{.eannot} files, as described
\href{https://zzz.bwh.harvard.edu/luna/ref/annotations/}{here})}
}
\value{
No explicit return value. Similar to \code{\link{lattach}}, which
is just a wrapper around the \code{\link{ledf}} function
}
\description{
Directly attaches an EDF
}
\examples{
\dontrun{
> ledf("edfs/learn-nsrr02.edf", "nsrr02", "edfs/learn-nsrr02-profusion.xml")
nsrr02 : 14 signals, 10 annotations, of 09:57:30 duration
}
}
