% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{leval.project}
\alias{leval.project}
\title{Evaluates an arbitrary set of Luna commands for all members of a project}
\usage{
leval.project(sl, command)
}
\arguments{
\item{sl}{a required sample-list, as returned by \code{\link{lsl}()}}

\item{command}{a single string or character vector of
\href{https://zzz.bwh.harvard.edu/luna/ref/}{Luna commands},
or the return value of \code{\link{lcmd}()}}
}
\value{
the same list-of-data-frames as returned by \code{\link{leval}()},
except each data-frame will now contain all individuals in the project
specified by sample list \code{sl}.
}
\description{
Evaluates a Luna commands for all individuals in a sample-list
}
\note{
Unlike \code{\link{leval}()}, the \code{leval.project()} function attaches
and then drops each EDF in the sample list, iteratively. This means that any
changes to the internal EDFs will not persist across different runs of
\code{leval.project()} in the way they do for \code{\link{leval}()}.
}
\examples{
\dontrun{
> k <- leval.project(lsl("s.lst") , lcmd("cmd.txt"))
## this does the same as the lunaC command
## \% luna s.lst -o out.db < cmd.txt
}

}
