% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lfilter}
\alias{lfilter}
\title{FIR filtering}
\usage{
lfilter(x, sr, lwr, upr, tw = 1, ripple = 0.02)
}
\arguments{
\item{x}{a time series vector}

\item{sr}{Sample rate of \code{x}}

\item{lwr}{Lower transition frequency}

\item{upr}{Upper transition frequency}
}
\value{
a filtered version of \code{x}
}
\description{
Applies a FIR bandpass filter
}
\note{
This is a wrapper around the method implemented by the \code{FILTER} command.
}
