% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lheatmap}
\alias{lheatmap}
\title{LUNA plotting}
\usage{
lheatmap(
  x,
  y,
  z,
  col = luna.globals$turbo.colors(100),
  mt = "",
  f = rep(T, length(z)),
  zero = rep(F, length(z)),
  xlines = NULL,
  ylines = NULL,
  zlim = NULL,
  win = NULL,
  legend = NULL
)
}
\arguments{
\item{x}{x-axis values}

\item{y}{y-axis values, of \code{length(x)}}

\item{z}{z-axis values, of \code{length(x)}}

\item{col}{100-valued color palette, defaults to \code{\link{lturbo}(100)}}

\item{mt}{main title (string)}

\item{f}{filter: boolean values of length equal to \code{length(x)}}

\item{zero}{boolean vector of length equal to \code{length(x)}; if T, set Z to 0}

\item{xlines}{draw vertical lines along the x-axis at these values (default: none)}

\item{ylines}{draw horizontal lines along the y-axus at these values (default: none)}

\item{zlim}{set range for Z axis (defaults to observed range in Z)}

\item{win}{winsorize Z at this percentile, e.g. 0.05}

\item{legend}{integer: return Z quantiles at this many uniformly-spaced positions}
}
\value{
plot is generated in the current graphics device; no return value unless legend is set
}
\description{
draws a rectangular heatmap (e.g. spectrogram)
}
\examples{
\dontrun{
k <- ldb("out.db")
d <- k$PSD$CH_E_F
lheatmap(d$E, d$F, 10*log10(d$PSD))
}
}
