% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lid}
\alias{lid}
\title{Extract a particular individual from a returned data frame}
\usage{
lid(d, id)
}
\arguments{
\item{d}{a data-frame in the format returned by \code{\link{lx}()}}

\item{id}{a vector of one or more string \code{ID} of the individual(s) to filter}
}
\value{
the row-subset of the data-frame for individual(s) in \code{id}
}
\description{
Extracts one or more particular individual(s) from a data-frame returned by \code{\link{lx}()}
}
\examples{
\dontrun{
> lid(lx(k, "EPOCH"), "nsrr02")
      ID DUR INC   NE
2 nsrr02  30  30 1195

> lid(lx(k, "EPOCH"), c("nsrr02", "nsrr03"))
      ID DUR INC   NE
2 nsrr02  30  30 1195
3 nsrr03  30  30 1364
}
}
