% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{llog}
\alias{llog}
\title{Turns Luna's typical console messages on/off}
\usage{
llog(x)
}
\arguments{
\item{x}{either \code{1} or \code{0} to turn console messaging on or off, respectively}
}
\description{
Turns on or off Luna's console message
}
\note{
By default, lunaR evaluates Luna commands more or less "silently". By
running \code{llog(1)}, you'll get more output, which can be useful to diagnose
errors, or check the progress of long-running jobs.
}
