% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_log.R
\name{log_add}
\alias{log_add}
\title{Add event to log}
\usage{
log_add(
  event = c("action", "error", "plot"),
  action = "modelfit",
  id,
  context = NA
)
}
\arguments{
\item{event}{type of event (\code{action} or \code{error})}

\item{action}{action type, e.g. \code{modelfit}}

\item{id}{run id}

\item{context}{other information, e.g. error message}
}
\description{
Add event to log
}
