% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lplasma}
\alias{lplasma}
\title{Generating a plasma palette (based on \code{viridis} package)}
\usage{
lplasma(n = 100)
}
\arguments{
\item{n}{number of colors to return}
}
\value{
vector of color values, length \code{n}
}
\description{
Generating a plasma palette (based on \code{viridis} package)
}
