% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lrefresh}
\alias{lrefresh}
\title{Reverts to the original attached EDF}
\usage{
lrefresh()
}
\value{
no explicit return value
}
\description{
Reattaches the currently attached EDF
}
\note{
When \code{lrefresh()}-ing an attached EDF, any previous in-memory modifications
(i.e. from masking, filtering, or other manipulations) are effectively reset.
}
