% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lreset}
\alias{lreset}
\title{Clears all previously lset() variables}
\usage{
lreset()
}
\description{
Clears all previously lset() variables
}
\note{
This function clears all variables (both user-defined and special variables).
It does not effect any currently attached EDF however. (In contrast, \code{\link{ldrop}()}
or \code{\link{lrefresh}()} will alter the currently-attached EDF, but do not
effect any variable definitions.)
}
