% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lsanitize}
\alias{lsanitize}
\title{Cleans syntax of command, factor/level and channel names}
\usage{
lsanitize(s)
}
\arguments{
\item{s}{a string to be cleaned}
}
\value{
a cleaned string
}
\description{
Replaces potentially troublesome characters with an underscore
}
\note{
This function is a wrapper for the R command \code{gsub("[^[:alnum:]]", "_", s)},
which replaces all non-alpha-numeric characters with an underscore. See the
Luna lunaR \href{https://zzz.bwh.harvard.edu/luna/tut/tut4/}{tutorial} for a motivating example.
}
\examples{
lsanitize( "C3/M1" )

}
