% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lset}
\alias{lset}
\title{Setting variables}
\usage{
lset(var, val = NULL)
}
\arguments{
\item{var}{a required parameter that is either a
\href{https://zzz.bwh.harvard.edu/luna/luna/args/#variables}{variable} name
(if \code{val} is non-\code{NULL}), or a list of variable/value pairs, or a
filename of a \href{https://zzz.bwh.harvard.edu/luna/luna/args/#parameter-files}{parameter file}}

\item{val}{if \code{val} is non-\code{NULL}, then \code{var} is interpreted
as the variable name, which is set to the value of \code{val}; if \code{val}
equals "\code{.}" this erases the variable \code{var}}
}
\description{
Sets a (special) variable, similar to lunaC command-line options
}
\note{
The special variables \code{sig} and \code{alias} operate slightly differently with \code{lset()},
in that new values are appended to the list, rather than over-writing the existing value.

\preformatted{
> lset("sig", "C3")
> lset("sig", "C4")
}
is the same as
\preformatted{
> lset("sig", "C3, C4")
}

To reset any variable (including \code{sig} and \code{alias})
\preformatted{
> lset("sig", ".")
}

the \code{\link{lreset}()} function clears all variables, and resets all
special variables to their default values.
}
\examples{
\dontrun{
## these are equivalent
## \% luna s.lst stage=N2 excl=arousal_standard sig=C3,C4 < cmd.txt

> lset("stage", "N2")
> lset("excl", "arousal_standard")
> lset("sig", "C3, C4")
> k <- leval(lcmd("cmd.txt"))

> lset(list(stage="N2", excl="arousal_standard", sig="C3, C4"))

## if these variables are specified in a file
> lset("files/param.txt")#'
}

}
