% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lsl}
\alias{lsl}
\title{Loads a sample list}
\usage{
lsl(file, path = "")
}
\arguments{
\item{file}{a required argument, giving the name of the sample-list file}

\item{path}{an optional argument, mirroring Luna's
\href{https://zzz.bwh.harvard.edu/luna/luna/args/#search-paths}{path} command-line option}
}
\value{
a named-list representing the sample-list
}
\description{
Imports a Luna \href{https://zzz.bwh.harvard.edu/luna/luna/args/#sample-lists}{sample-list}
into R
}
\examples{
\dontrun{
> sl <- lsl("s.lst")
3 observations in s.lst

> names(sl)
[1] "nsrr01" "nsrr02" "nsrr03"

> str(sl)
List of 3
$ nsrr01:List of 2
..$ EDF  : chr "edfs/learn-nsrr01.edf"
..$ ANNOT: chr "edfs/learn-nsrr01-profusion.xml"
$ nsrr02:List of 2
..$ EDF  : chr "edfs/learn-nsrr02.edf"
..$ ANNOT: chr "edfs/learn-nsrr02-profusion.xml"
$ nsrr03:List of 2
..$ EDF  : chr "edfs/learn-nsrr03.edf"
..$ ANNOT: chr "edfs/learn-nsrr03-profusion.xml"
}
}
