% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lstat}
\alias{lstat}
\title{report on the in-memory EDF what is in memory}
\usage{
lstat()
}
\value{
no explicit return values, other than output to console
}
\description{
One line description to console of the currently-attached EDF
}
\note{
This command is automatically called after each \code{\link{leval}} or
\code{\link{lattach}} command.
}
