% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lstgcols}
\alias{lstgcols}
\title{Maps typical stage labels to colors (for plotting)}
\usage{
lstgcols(s)
}
\arguments{
\item{s}{a required parameter, a vector of stage label, that should be in the
format \code{NREM1, NREM2, NREM3, NREM4, REM} and \code{wake}.}
}
\value{
a corresponding vector of colors
}
\description{
Convenience function to return colors for stage labels, to be used in plotting
}
\examples{
lstgcols(c("NREM1", "NREM2", "NREM3", "NREM4", "REM", "wake", "?"))
}
