% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ltopo.dconn}
\alias{ltopo.dconn}
\title{LUNA plotting}
\usage{
ltopo.dconn(
  ch,
  chs1,
  chs2,
  z,
  flt = T,
  zr = NULL,
  cex = 2,
  w = 8,
  title = "",
  head = T,
  signed = F
)
}
\arguments{
\item{ch}{single character value: channel to seed on}

\item{chs1}{channel 'a' as per \code{\link{ltopo.conn}()}}

\item{chs2}{channel 'b' as per \code{\link{ltopo.conn}()}}

\item{z}{numeric vector of pairwise Z values}

\item{flt}{optional boolean vector of length(z), to filter observations in/out of the plot}

\item{zr}{range of Z axis}

\item{cex}{size of channel circles}

\item{w}{width of lines between channels}

\item{title}{main title text}

\item{head}{boolean: draw simple head circle if T}

\item{signed}{assume that Z is directional (see \code{\link{ltopo.conn}()} notes)}
}
\value{
plot is generated in the current graphics device; no return value
}
\description{
draw	a connectivity topoplot	(i.e. of lines between pairs of	channels, but seeding on one channel,
i.e. as a special case of \code{\link{ltopo.conn}()}
}
\note{
This is a special case of \code{\link{ltopo.conn}()}, where instead of showing all available pairwise measures,
only those where channel A matches the single channel specified by \code{ch} are shown.
}
\examples{
\dontrun{
ltopo.dconn(ch = "C3", chs1 = coh$CH1, chs2 = coh$CH2, z = coh$COH, flt = coh$B == "SIGMA")
}

}
