% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ltopo.heat}
\alias{ltopo.heat}
\title{LUNA plotting}
\usage{
ltopo.heat(
  c,
  z,
  sz = 1,
  flt = rep(T, length(z)),
  zlab = "",
  mt = "",
  zlim = NULL,
  th = NA,
  th.z = z,
  show.leg = F,
  zeroed = F,
  head = F
)
}
\arguments{
\item{c}{vector of channel labels}

\item{z}{vector of values to plot (same length as \code{c})}

\item{sz}{integer: size of channel circles plotted (default = 1)}

\item{flt}{boolean filter of length equal to \code{length(c)}}

\item{zlab}{label for Z-axis}

\item{mt}{main title label}

\item{zlim}{set Z range (defaults to observed range)}

\item{th}{single numeric value threshold for highlighting certain channels (default NULL)}

\item{th.z}{vector, same length as \code{c} and \code{z}; channels with \code{th.z} \eqn{\ge} \code{th} are highlighted}

\item{show.leg}{boolean: plot a legend (default T)}

\item{zeroed}{boolean: make Z range symmetric around zero if T (default = F)}

\item{head}{boolean: draw simple head image (circle, default = F)}
}
\value{
plot is generated in the current graphics device; no return value
}
\description{
draws a 'topoplot' using a heatmap color scale
}
\examples{
\dontrun{
ltopo.heat(psd$CH, psd$PSD, flt = psd$F == 15, sz = 3, zlab = "Power 15 Hz", mt = "Plot 1")
}
}
