% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ltopo.rb}
\alias{ltopo.rb}
\title{LUNA plotting}
\usage{
ltopo.rb(
  c,
  z,
  flt = rep(T, length(z)),
  sz = 1,
  zlab = "",
  mt = "",
  zlim = NULL,
  th = NULL,
  th.z = z,
  show.leg = F,
  zeroed = T,
  head = F,
  col = colorRampPalette(c("blue", "white", "red"))(101),
  ring.lwd = 1
)
}
\arguments{
\item{c}{vector of channel labels}

\item{z}{vector of values to plot (same length as \code{c})}

\item{flt}{boolean filter of length equal to \code{length(c)}}

\item{sz}{integer: size of channel circles plotted (default = 1)}

\item{zlab}{label for Z-axis}

\item{mt}{main title label}

\item{zlim}{set Z range (defaults to observed range)}

\item{th}{single numeric value threshold for highlighting certain channels (default NULL)}

\item{th.z}{vector, same length as \code{c} and \code{z}; channels with \code{th.z} \eqn{\ge} \code{th} are highlighted}

\item{show.leg}{boolean: plot a legend (default T)}

\item{zeroed}{boolean: make Z range symmetric around zero if T (default = F)}

\item{head}{boolean: draw simple head image (circle, default = F)}

\item{col}{101-valued palette, defaults to \code{colorRampPalette(c("blue", "white", "red"))(101)}}

\item{ring.lwd}{width of rings around highlighted channels (defaults to 1)#'}
}
\value{
plot is generated in the current graphics device; no return value
}
\description{
draws a 'topoplot' using a red-blue color scale
}
\note{
\code{ltopo.rb(...)} is the same as \code{\link{ltopo.heat}(...)} with
\code{zeroed = T} and a different color scheme (blue as negative, red as positive, white as 0)
}
