% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{ltopo.topo}
\alias{ltopo.topo}
\title{LUNA plotting}
\usage{
ltopo.topo(
  c,
  c2,
  z,
  zlim = NULL,
  f = rep(T, length(z)),
  sz = 0.05,
  sz2 = 0.05,
  ring.lwd = 1,
  same.cols = T,
  col = rbpal,
  zlab = "",
  mt = "",
  zeroed = T
)
}
\arguments{
\item{c}{character vector of channel labels for the inner plots}

\item{c2}{character vector of channel labels for the outer plots}

\item{z}{numeric vector of values to plot (by Z-axis color scale)}

\item{zlim}{set Z range (defaults to observed range)}

\item{f}{boolean vector: filter of \code{length(z)}}

\item{sz}{relative size of each inner point (default 0.05)}

\item{sz2}{relative size of each outer plot (default 0.05)}

\item{ring.lwd}{width of ring around each point}

\item{same.cols}{use the same Z color range for each inner topoplot}

\item{col}{optional color palette (100-elements)}

\item{zlab}{optional Z-axis label for legend}

\item{mt}{optional main title text (default "")}

\item{zeroed}{boolean value: if T, set Z ranges to be symmetric around 0}
}
\value{
plot is generated in the current graphics device; no return value
}
\description{
draw a 'topoplot of topoplots'
}
\examples{
\dontrun{
ltopo.topo(c = c(coh$CH1, coh$CH2), c2 = c(coh$CH2, coh$CH1),
    z = c(coh$ICOH, -1 * coh$ICOH), f = rep(coh$B == "SIGMA", 2),
    sz=0.08, sz2=0.6)
}

}
