% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lturbo}
\alias{lturbo}
\title{Generating a turbo palette}
\usage{
lturbo(n = 100)
}
\arguments{
\item{n}{number of colors to return}
}
\value{
vector of color values, length \code{n}
}
\description{
Generating a turbo palette
}
