% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_misc.R
\name{luna_cache_get}
\alias{luna_cache_get}
\title{Get element from active project
Just a convenience wrapper with error checking}
\usage{
luna_cache_get(entry = c("name", "folder"))
}
\description{
Get element from active project
Just a convenience wrapper with error checking
}
