% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_clone.R
\name{luna_clone}
\alias{luna_clone}
\title{Clone model}
\usage{
luna_clone(id, new_id, folder = NULL, force = FALSE, verbose = TRUE)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{new_id}{new run id}

\item{folder}{path to folder containing the model file. Default is current directory.}
}
\description{
Clone model
}
