% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_compare.R
\name{luna_compare}
\alias{luna_compare}
\title{Compare basic run info for 2 or more model runs}
\usage{
luna_compare(..., folder = ".", return_object = FALSE)
}
\arguments{
\item{...}{id of runs to compare. Can be specified either
as a vector, or as separate arguments.}

\item{folder}{folder to look in}

\item{return_object}{return object?}
}
\description{
Compare basic run info for 2 or more model runs
}
