% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_config.R
\name{luna_config}
\alias{luna_config}
\title{Edit luna config file in editor, e.g. RStudio}
\usage{
luna_config(verbose = TRUE)
}
\arguments{
\item{verbose}{verbose output}
}
\description{
Edit luna config file in editor, e.g. RStudio
}
