% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_dataset.R
\name{luna_dataset}
\alias{luna_dataset}
\title{Get data from a model}
\usage{
luna_dataset(id, folder = NULL)
}
\arguments{
\item{id}{run id, e.g. \code{run1}}

\item{folder}{folder containing the model file}
}
\description{
Get data from a model
}
