% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_diff.R
\name{luna_diff}
\alias{luna_diff}
\title{Show a diff between two models}
\usage{
luna_diff(id, reference = NULL, folder = NULL, verbose = TRUE)
}
\arguments{
\item{id}{either a single \code{id} or a vector of exactly two model \code{id}s.
If specified as a single value, the \code{id} needs to have a \code{reference} model
specified to which it can be compared. If \code{id} is of length 2, then it will
just diff the models for the two \code{id}s.}

\item{reference}{id for reference model to compare to (optional)}

\item{folder}{path to folder containing the model file. Default is current directory.}
}
\description{
Show a diff between two models
}
