% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_edit.R
\name{luna_edit}
\alias{luna_edit}
\title{Edit model in editor, e.g. RStudio}
\usage{
luna_edit(id, folder = NULL, verbose = TRUE)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{folder}{path to folder containing the model file. Default is current directory.}
}
\description{
Edit model in editor, e.g. RStudio
}
