% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_edit_project.R
\name{luna_edit_project}
\alias{luna_edit_project}
\title{Edit luna config file in editor, e.g. RStudio}
\usage{
luna_edit_project(verbose = TRUE)
}
\arguments{
\item{verbose}{verbose output}
}
\description{
Edit luna config file in editor, e.g. RStudio
}
