% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_gof.R
\name{luna_gof}
\alias{luna_gof}
\title{Creates a ggplot2 panel object with basic goodness of fit plots}
\usage{
luna_gof(
  id,
  folder = NULL,
  residual = c("CWRES", "NPDE"),
  theme = ggplot2::theme_classic,
  smooth_method = "loess",
  ltbs = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the model
is run.}

\item{folder}{path to folder containing the model file. Default is current directory.}

\item{residual}{either "CWRES" or "NPDE"}

\item{smooth_method}{ggplot2-supported smooth method, e.g. "loess"}

\item{ltbs}{log-transform-both-sides error model? If \code{TRUE}, will
exponentiate DV, PRED, and IPRED.}
}
\description{
Creates a ggplot2 panel object with basic goodness of fit plots
}
