% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_help.R
\name{luna_help}
\alias{luna_help}
\title{Ask AI agent to help}
\usage{
luna_help(prompt = NULL, model = "gpt-4.1-mini", verbose = TRUE, ...)
}
\arguments{
\item{prompt}{question to ask LLM. By default NULL, which means it will
look at the log and extract the last event, e.g. an error from NONMEM}

\item{model}{LLM, e.g. "gpt-4.1-mini"}

\item{verbose}{verbose output?}

\item{...}{optional parameters passed onto ellmer::chat()}
}
\description{
By default it will pick the most recent error from NONMEM
as prompt, but can also ask questions.
}
\examples{
\dontrun{
luna_help() # will get help from LLM on last NONMEM error
luna_help("How do I write a model with non-linear elimination?")
}

}
