% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_ind.R
\name{luna_ind}
\alias{luna_ind}
\title{Creates a ggplot2 plot with predictions and observations, split
by subject}
\usage{
luna_ind(
  id,
  folder = NULL,
  theme = ggplot2::theme_classic,
  smooth_method = "loess",
  ltbs = FALSE,
  ncol = 3,
  nrow = 3,
  page = 1,
  scales = NULL,
  subject_ids = NULL,
  verbose = TRUE
)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{folder}{path to folder containing the model file. Default is current directory.}

\item{smooth_method}{ggplot2-supported smooth method, e.g. "loess"}

\item{ltbs}{log-transform-both-sides error model? If \code{TRUE}, will
exponentiate DV, PRED, and IPRED.}

\item{ncol}{number of faceting columns to show}

\item{nrow}{number of faceting rows to show}

\item{page}{page number of plots to show}

\item{scales}{passed to \code{facet_wrap()} function as the \code{scales}
argument, can be "free", "free_x", or "free_y".}

\item{subject_ids}{list of subject IDs to be plotted.}
}
\description{
Creates a ggplot2 plot with predictions and observations, split
by subject
}
