% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_info.R
\name{luna_info}
\alias{luna_info}
\title{Show info for a model run}
\usage{
luna_info(id, folder = ".", return_object = FALSE, ...)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{folder}{path to folder containing the model file. Default is current directory.}
}
\description{
Show info for a model run
}
