% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_list.R
\name{luna_list}
\alias{luna_list}
\title{List runs in current project}
\usage{
luna_list(raw = FALSE)
}
\arguments{
\item{raw}{if \code{TRUE}, will return a tibble with all information
this is more useful for exporting to a file rather than viewing
in console.}
}
\description{
List runs in current project
}
