% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_list_models.R
\name{luna_list_models}
\alias{luna_list_models}
\title{Show a listing of finished, running, and to-be-run models}
\usage{
luna_list_models(
  folder = ".",
  project = NULL,
  verbose = FALSE,
  as_data = FALSE
)
}
\description{
Show a listing of finished, running, and to-be-run models
}
