% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_list_runs.R
\name{luna_list_runs}
\alias{luna_list_runs}
\title{Show a listing of finished, running, and to-be-run models}
\usage{
luna_list_runs(
  folder = ".",
  project = NULL,
  arrange = c("type", "status", "id"),
  verbose = FALSE,
  format = "pipe"
)
}
\description{
Show a listing of finished, running, and to-be-run models
}
