% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_load_project.R
\name{luna_load_project}
\alias{luna_load_project}
\title{Load a project from yaml and gather results}
\usage{
luna_load_project(name = NULL, folder = ".", verbose = TRUE)
}
\arguments{
\item{name}{name of the project, will be used as as the filename for the
YAML file. Cannot contain spaces.}

\item{folder}{folder to create the project in. If the folder doesn't exist,
it will be created.}

\item{verbose}{verbosity}
}
\description{
Load a project from yaml and gather results
}
