% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_log.R
\name{luna_log}
\alias{luna_log}
\title{Show a log of events (actions and errors)}
\usage{
luna_log(n = 20)
}
\arguments{
\item{n}{number of most recent events to return, default is 20}
}
\description{
Show a log of events (actions and errors)
}
