% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_note.R
\name{luna_note}
\alias{luna_note}
\title{Annotate a run
Either add a note or clear notes from a run}
\usage{
luna_note(id, note = NULL, clear = FALSE, force = TRUE, verbose = TRUE)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{note}{note(s) to add (character vector)}

\item{clear}{should existing notes be cleared? Default is \code{FALSE}}
}
\description{
Annotate a run
Either add a note or clear notes from a run
}
\examples{
\dontrun{
luna_note("run1", "Initial model") # adds a note
luna_note("run1", "Initial model", clear=TRUE) # first clears existing notes, then adds a note
luna_note("run1", clear=TRUE) # only clears existing notes
}

}
