% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_run.R
\name{luna_run_ferx}
\alias{luna_run_ferx}
\title{Run a ferx model (internal helper for luna_run)}
\usage{
luna_run_ferx(id, folder, config, ...)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the model
is run.}

\item{folder}{path to folder containing the model file. Default is current directory.}

\item{config}{resolved luna config list}
}
\description{
Run a ferx model (internal helper for luna_run)
}
\keyword{internal}
