% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_sync.R
\name{luna_sync}
\alias{luna_sync}
\title{Synchronize (re-load) luna project}
\usage{
luna_sync(...)
}
\description{
This function is essentially a duplicate of luna_load_project()
at the moment. It may take on different roles in the future.
}
