% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_tables.R
\name{luna_tables}
\alias{luna_tables}
\title{Get tables from a model / run}
\usage{
luna_tables(id, folder = NULL, verbose = TRUE)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{folder}{path to folder containing the model file. Default is current directory.}
}
\description{
Get tables from a model / run
}
