% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_tool.R
\name{luna_tool}
\alias{luna_tool}
\title{Run external tools (e.g. diagnostics) on models or run outputs
This can be used e.g. for bootstraps and VPCs. The function is implemented
in a modular way so that it can be easily extended.}
\usage{
luna_tool(id, tool = NULL, force = FALSE, as_job = NULL, verbose = TRUE)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{tool}{id for the tool, needs to be referenced in project YAML. See examples
for further details.}

\item{as_job}{run as an RStudio job (async), or in the console. If left \code{NULL}
will use setting in luna config.}

\item{verbose}{verbose output}
}
\description{
Run external tools (e.g. diagnostics) on models or run outputs
This can be used e.g. for bootstraps and VPCs. The function is implemented
in a modular way so that it can be easily extended.
}
