% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_xpose.R
\name{luna_xpose}
\alias{luna_xpose}
\title{Read tables for a model and parse into an Xpose database}
\usage{
luna_xpose(id, folder = NULL, verbose = TRUE)
}
\arguments{
\item{id}{run id}

\item{folder}{folder}

\item{verbose}{verbose output?}
}
\description{
Read tables for a model and parse into an Xpose database
}
