% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lvar}
\alias{lvar}
\title{Getting variables}
\usage{
lvar(v)
}
\arguments{
\item{v}{a required parameter, the variable name}
}
\value{
the variable value, or \code{NULL} if the variable is not defined, or
if it is a special variable (e.g. \code{sig} or \code{alias})
}
\description{
Returns the value of a user-defined variable
}
\note{
The \code{lvar()} function only applies to user-defined variables such as \code{${stage}}.
Special variables (e.g. \code{sig, annots}, etc) will currently return \code{NULL} here.
\preformatted{
> lvar("sig")
NULL
}
This doesn't mean they are not defined. Whilst not a bug exactly, this "feature"
is not particularly convenient; future releases of \emph{lunaR} will be fixed to return
the actual values of all variables.
}
\examples{
\dontrun{
> lvar("stage")
[1] "N2"
}

}
