% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lwin}
\alias{lwin}
\title{Winsorize a vector}
\usage{
lwin(x, p = 0.05)
}
\arguments{
\item{x}{a numeric vector}

\item{p}{percentile}
}
\value{
winsorized version of \code{x}
}
\description{
Winsorize a vector
}
\examples{
dcd
dc

}
