% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{lx2}
\alias{lx2}
\title{Like lx() but for lists of returned objects}
\usage{
lx2(k, ...)
}
\arguments{
\item{k}{multiple returned values from \code{\link{leval}()} or similar have
been assembled as a list of lists}

\item{...}{all parameters used in \code{\link{lx}()}, including the command
and strata specification}
}
\value{
same output as \code{\link{lx}()}
}
\description{
A version of lx() designed to work with lists of returned lists
}
\examples{
\dontrun{
> k <- list()
> k[[ "S1" ]] <- leval( some-commands-go-here )
> k[[ "S2" ]] <- leval( some-more-commands-go-here )

> lx2(k)
}
}
