% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonlight.R
\name{moonlight}
\alias{moonlight}
\title{Initiate the Moonlight viewer in a browser window}
\usage{
moonlight(
  sample.list = NULL,
  proj.path = NULL,
  nap.mode = FALSE,
  environ.file = NULL,
  local = T
)
}
\arguments{
\item{sample.list}{sample list (defaults to \code{s.lst})}

\item{proj.path}{working folder for Moonlight (default current)}

\item{nap.mode}{boolean value to indicate whether to expect NAP output (default F)}

\item{environ.file}{if non-NULL, specify a file with environment variables}

\item{local}{boolean value to indicate whether running locally (versus Docker container) (default T)}
}
\description{
Initiate the Moonlight viewer in a browser window
}
