% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_new_project.R
\name{new_project_from_template}
\alias{new_project_from_template}
\title{Create a new project object based on a predefined template}
\usage{
new_project_from_template(template = "default")
}
\arguments{
\item{template}{template for new project. Either a reference to one of the
built-in templates (e.g. \code{default}), or path to a file.}
}
\description{
Create a new project object based on a predefined template
}
