% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoc.R
\name{outliers.inc}
\alias{outliers.inc}
\title{Filter outliers with masking}
\usage{
outliers.inc(
  x,
  inc = rep(T, length(x)),
  m = mean(x[inc], na.rm = T),
  sdev = sd(x[inc], na.rm = T),
  t = 3
)
}
\arguments{
\item{x}{numeric vector}

\item{inc}{boolean vector with same length as \code{x}, used for masking.
\code{FALSE} will set the element to \code{NA} in the corresponding position
in \code{x}.}

\item{m}{target center of acceptable interval after outlier removal,
default to \code{x}'s mean}

\item{sdev}{target variation of acceptable interval, default to \code{x}'s
standard deviation}

\item{t}{scale of deviation, default to 3}
}
\value{
\code{x} after setting outliers to \code{NA}
}
\description{
Set outliers in a vector to \code{NA}
}
