% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoc.R
\name{outliers}
\alias{outliers}
\title{Filter outliers}
\usage{
outliers(x, m = mean(x, na.rm = T), sdev = sd(x, na.rm = T), t = 3)
}
\arguments{
\item{x}{numeric vector}

\item{m}{target center of acceptable interval after outlier removal,
default to \code{x}'s mean}

\item{sdev}{target variation of acceptable interval, default to \code{x}'s
standard deviation}

\item{t}{scale of deviation, default to 3}
}
\value{
\code{x} after setting outliers to \code{NA}
}
\description{
Set outliers in a vector to \code{NA}
}
