% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{parse_psn_options}
\alias{parse_psn_options}
\title{Parse PsN arguments specified as a list into a vector of command
line arguments}
\usage{
parse_psn_options(x)
}
\arguments{
\item{x}{list of arguments and values, e.g. \code{list(samples=100, threads=2)}}
}
\description{
Parse PsN arguments specified as a list into a vector of command
line arguments
}
