% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pluck}
\alias{pluck}
\title{Pluck an inner list element from an outer list, where an element matches
(default name is "id") a specific value. Analogue to _.pluck() in
JS/underscore}
\usage{
pluck(x, id, el = "id")
}
\description{
Pluck an inner list element from an outer list, where an element matches
(default name is "id") a specific value. Analogue to _.pluck() in
JS/underscore
}
