% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.luna.project}
\alias{print.luna.project}
\title{Prints a list of models / runs in project}
\usage{
\method{print}{luna.project}(x, arrange = c("type", "status", "id"), format = "simple", ...)
}
\arguments{
\item{project}{luna project object}
}
\description{
Prints a list of models / runs in project
}
