% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.luna.run_table}
\alias{print.luna.run_table}
\title{Prints a list of models / runs in project}
\usage{
\method{print}{luna.run_table}(x, format = "simple", verbose = FALSE, ...)
}
\arguments{
\item{x}{luna project object}
}
\description{
Prints a list of models / runs in project
}
