% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nlme.R
\name{print_nmfe_output}
\alias{print_nmfe_output}
\title{Print nmfe output (stdout and stderr) from a run folder}
\usage{
print_nmfe_output(path, results_file = "run.lst")
}
\arguments{
\item{path}{path to folder with NMFE run}

\item{results_file}{name of output file}
}
\description{
Print nmfe output (stdout and stderr) from a run folder
}
