% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tables_from_model.R
\name{remove_tables_from_model}
\alias{remove_tables_from_model}
\title{Remove all $TABLE records from a model}
\usage{
remove_tables_from_model(model)
}
\arguments{
\item{model}{pharmpy model object}
}
\description{
Remove all $TABLE records from a model
}
