% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{replace_list_elements}
\alias{replace_list_elements}
\title{Replace elements in a list, based on elements in a second list}
\usage{
replace_list_elements(x, y)
}
\description{
Replace elements in a list, based on elements in a second list
}
