% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nlme.R
\name{run_nlme}
\alias{run_nlme}
\title{Run model in NONMEM}
\usage{
run_nlme(
  model,
  data = NULL,
  id,
  path = getwd(),
  method = c("nmfe", "pharmpy", "psn"),
  nmfe = get_nmfe_location_for_run(),
  force = FALSE,
  console = TRUE,
  save_fit = TRUE,
  save_summary = TRUE,
  auto_stack_encounters = TRUE,
  clean = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{model}{pharmpy model object or NONMEM model code (character) or path
to NONMEM model file.}

\item{data}{dataset (data.frame). Optional, can also be included in \code{model}
object (if specified as pharmpy model object).}

\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run. If no folder is specified, it will create a folder \code{run1} in
the current working directory, and will increment the run number for each
subsequent run.}

\item{path}{path to nonmem model. If not specified, will assume current
working directory.}

\item{method}{run method, either \code{pharmpy} dispatch, \code{nmfe} or \code{psn}
(psn::execute).}

\item{nmfe}{full path to nmfe file to run NONMEM with, if \code{method=="nmfe"}.}

\item{force}{if run folder (\code{id}) exists, should existing results be
removed before rerunning NONMEM? Default \code{FALSE}.}

\item{console}{show stderr and stdout in R console? If FALSE, will stream
to files \code{stdout} and \code{stderr} in fit folder.}

\item{save_fit}{save fit object. If \code{TRUE}, will save as <run_id.rds>. Can
also specify filename (rds) to save to.}

\item{save_summary}{save fit summary and parameter estimates to file?
Default is \code{TRUE}. Will use current folder, and save as
\verb{fit_summary_<id>.txt} and \verb{fit_parameters_<id>.csv}.}

\item{auto_stack_encounters}{only invoked if \code{data} argument supplied, not if
a pharmpy model object is supplied without \code{data}.
Detects if TIME within an individual is
decreasing from one record to another, which NONMEM cannot handle.
If this happens, it will add a reset event (EVID=3) at that time, and
increase the TIME for subsequent events so that NONMEM does not throw an
error. It will increase the time for the next encounter to the maximum
encounter length across all subjects in the dataset (rounded up to 100).
If no decreasing TIME is detected, nothing will be done (most common case).
This feature is useful e.g. for crossover trials when data on the same
individual ispresent but is included in the dataset as time-after-dose and
not actual time since first overall dose.}

\item{clean}{clean up run folder after NONMEM execution?}

\item{verbose}{verbose output?}
}
\description{
Run the model directly using nmfe (not through pharmpy).
This is a more reliable way of running NONMEM, and it is now possible to
stream stdout and stderr to file or to console, which is useful for
inspection of intermediate model fit.
}
\details{
The function does take a pharmpy model as input (optionally), and uses
pharmpy to read the results from the model fit, and returns a pharmpy
\code{modelfit} object.
}
