% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pharmpy_tool.R
\name{run_pharmpy_tool}
\alias{run_pharmpy_tool}
\title{Run a pharmpy tool, like bootstrap, or modelsearch}
\usage{
run_pharmpy_tool(id, model, results, tool = "bootstrap", verbose = FALSE, ...)
}
\arguments{
\item{id}{model id. Optional. If not specified, will generate random modelfit
id. The \code{id} will be used to create the run folder.}

\item{model}{Pharmpy model object, preferably created using
\code{luna::create_model()}.}

\item{verbose}{verbose output?}
}
\value{
fit object
}
\description{
Run a pharmpy tool, like bootstrap, or modelsearch
}
