% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run simulations}
\usage{
run_sim(
  fit = NULL,
  data = NULL,
  model = NULL,
  id = "sim1",
  force = FALSE,
  t_obs = NULL,
  dictionary = list(ID = "ID", DV = "DV", EVID = "EVID", AMT = "AMT", RATE = "RATE", CMT
    = "CMT", MDV = "MDV"),
  method = c("full", "sample"),
  regimen = NULL,
  covariates = NULL,
  tool = c("auto", "nonmem", "nlmixr2"),
  n_subjects = NULL,
  n_iterations = 1,
  include_in_output = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{dataset (data.frame). Optional, can also be included in \code{model}
object (if specified as pharmpy model object).}

\item{model}{pharmpy model object or NONMEM model code (character) or path
to NONMEM model file.}

\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run. If no folder is specified, it will create a folder \code{run1} in
the current working directory, and will increment the run number for each
subsequent run.}

\item{force}{if run folder (\code{id}) exists, should existing results be
removed before rerunning NONMEM? Default \code{FALSE}.}

\item{t_obs}{a vector of observations times. If specified, will override
the observations in each subject in the input dataset.}

\item{method}{either \code{full} or \code{sample}, meaning either use the input dataset
used to fit the model as-is, or sample subjects from the input dataset. When
sampling, the \code{n_subjects} argument determines the number of subjects to
sample, wihch can be either lower or higher than the actual number in the
input dataset. Sampling will be done with replacement. Both the \code{regimen}
and \code{covariates} in the input dataset can be overriden by these respective
arguments, see documentation for these arguments.}

\item{regimen}{if specified, will replace the regimens for each subject with
a custom regimen, specified using arguments \code{dose}, \code{interval}, and \code{route}
(and \code{t_inf} / \code{rate} for infusions).
E.g. \code{regimen = list(dose = 500, interval = 12, route = "oral")}.}

\item{covariates}{if specified, will replace subjects with subjects specified
in a data.frame. In the data.frame, the column names should correspond
exactly to any covariates included in the model. An \code{ID} column is required,
and for time-varying covariates, a \code{TIME} column is also required (otherwise
it will be assumed covariates are not changing over time).}

\item{n_subjects}{number of subjects to simulate}

\item{n_iterations}{number of iterations of the entire simulation to
perform. The dataset for the simulation will stay the same between each
iterations.}

\item{verbose}{verbose output?}
}
\value{
data.frame with simulation results
}
\description{
Run simulations
}
