% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_new_project.R
\name{save_project}
\alias{save_project}
\title{Save a project to a YAML file}
\usage{
save_project(project, path, verbose = FALSE, force = FALSE)
}
\arguments{
\item{project}{uno project object}

\item{path}{path to file to save project}

\item{verbose}{verbosity}

\item{force}{force saving of the file, even if it already exists}
}
\description{
Save a project to a YAML file
}
