% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_compartment_scale.R
\name{set_compartment_scale}
\alias{set_compartment_scale}
\title{Set scaling for certain compartments, e.g. dose and observation
compartments.}
\usage{
set_compartment_scale(
  model,
  compartment = 1,
  expression = list(variable = "V", scale = 1000),
  verbose = TRUE
)
}
\arguments{
\item{model}{pharmpy model object or NONMEM model code (character) or path
to NONMEM model file.}

\item{compartment}{compartment number}

\item{expression}{specification of new scaling, should always contain variable
and scale arguments. E.g. \code{list(variable = "V", "scale" = 1000)}.}

\item{verbose}{verbose output?}
}
\value{
Pharmpy model
}
\description{
Currently not available in Pharmpy, this is a workaround function.
}
