\name{somma}
\alias{somma}
\title{Sum Two Numbers}
\description{
Computes the sum of two numeric values.
}
\usage{
somma(x, y)
}
\arguments{
\item{x}{A numeric value.}
\item{y}{A numeric value.}
}
\details{
The function takes two numeric inputs and returns their sum.
}
\value{
A numeric value equal to the sum of \code{x} and \code{y}.
}
\examples{
somma(2, 3)
somma(-1, 5)
}
\author{
Your Name
}
\seealso{
\code{\link{sum}}
}
\keyword{arith}
