% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{truncate_columns}
\alias{truncate_columns}
\title{Apply column width truncation based on console width and relative width specs}
\usage{
truncate_columns(df, width_specs, total_width = NULL)
}
\description{
Apply column width truncation based on console width and relative width specs
}
