% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cache.R
\name{update_cache}
\alias{update_cache}
\title{Update luna disk cache}
\usage{
update_cache(folder, name)
}
\arguments{
\item{folder}{folder to create the project in. If the folder doesn't exist,
it will be created.}

\item{name}{name of the project, will be used as as the filename for the
YAML file. Cannot contain spaces.}
}
\description{
Update luna disk cache
}
