% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_estimation_method.R
\name{update_estimation_method}
\alias{update_estimation_method}
\title{Wrapper around pharmr's functions to set/add estimation methods}
\usage{
update_estimation_method(model, estimation_method)
}
\arguments{
\item{model}{pharmpy model object or NONMEM model code (character) or path
to NONMEM model file.}

\item{estimation_method}{Optional. Character vector of estimation method(s)
to apply to model. Will remove all existing estimation steps in the model
and update with methods specified in argument.}
}
\description{
The current pharmpy functionatliy is not stable, hence the need for this
wrapper.
}
