% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_parameters.R
\name{update_parameters}
\alias{update_parameters}
\title{Update parameter estimates (and fix)}
\usage{
update_parameters(model, fit, fix = TRUE, verbose = TRUE)
}
\arguments{
\item{model}{pharmpy model object or NONMEM model code (character) or path
to NONMEM model file.}

\item{fit}{pharmpy fit object}

\item{fix}{fix the estimates? Currently this can only be \code{TRUE}, since this is the
only option that Pharmpy/pharmr supports.}

\item{verbose}{verbose output?}
}
\description{
For example for using model in simulations.
}
