% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoc.R
\name{zf}
\alias{zf}
\title{Vector normalization}
\usage{
zf(x, m = mean(x, na.rm = T), sdev = sd(x, na.rm = T))
}
\arguments{
\item{x}{numeric vector}

\item{m}{target center of normalization, default to \code{x}'s mean}

\item{sdev}{target deviation of normalization, default to \code{x}'s
standard deviation}
}
\value{
normalized vector with same length as \code{x}
}
\description{
Vector normalization
}
