% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_pharmpy_tool.R
\name{call_pharmpy_tool}
\alias{call_pharmpy_tool}
\title{Generic function for running a pharmpy tool, like bootstrap,
or modelsearch. A separate function is available for \code{fit()}}
\usage{
call_pharmpy_tool(id, model, results, tool = "bootstrap", verbose = FALSE, ...)
}
\arguments{
\item{id}{model id. Optional. If not specified, will generate random modelfit
id. The \code{id} will be used to create the run folder.}

\item{model}{Pharmpy model object, preferably created using
\code{luna::create_model()}.}

\item{verbose}{verbose output?}
}
\value{
fit object
}
\description{
Generic function for running a pharmpy tool, like bootstrap,
or modelsearch. A separate function is available for \code{fit()}
}
