% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cache.R
\name{create_cache}
\alias{create_cache}
\title{Create a cache environment to store settings for project and system}
\usage{
create_cache(project, verbose = FALSE)
}
\arguments{
\item{project}{luna project object}

\item{verbose}{verbose output?}
}
\description{
Create a cache environment to store settings for project and system
}
