% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_luna_cache.R
\name{create_cache}
\alias{create_cache}
\title{Create a cache environment to store settings for project and system}
\usage{
create_cache(name, folder, verbose = FALSE)
}
\arguments{
\item{name}{optional, name of project (matching to yaml file)}

\item{folder}{path to project}
}
\description{
Create a cache environment to store settings for project and system
}
