% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_check.R
\name{luna_check}
\alias{luna_check}
\title{Syntax-check a NONMEM model}
\usage{
luna_check(id, folder = NULL, verbose = FALSE, ...)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{folder}{path to folder containing the model file. Default is current directory.}
}
\description{
Syntax-check a NONMEM model
}
