% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_check.R
\name{luna_check}
\alias{luna_check}
\title{Syntax-check a NONMEM model}
\usage{
luna_check(id, folder = NULL, verbose = FALSE, ...)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run. If no folder is specified, it will create a folder \code{run1} in
the current working directory, and will increment the run number for each
subsequent run.}

\item{verbose}{verbose output?}
}
\description{
Syntax-check a NONMEM model
}
