% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_list.R
\name{luna_list}
\alias{luna_list}
\title{List runs in current project}
\usage{
luna_list(project = NULL)
}
\arguments{
\item{project}{project name. Default is \code{NULL}, will use active project}
}
\description{
List runs in current project
}
