% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_list.R
\name{luna_list}
\alias{luna_list}
\title{Show a listing of finished, running, and to-be-run models}
\usage{
luna_list(project = NULL, folder = ".", verbose = FALSE, as_data = FALSE)
}
\description{
Show a listing of finished, running, and to-be-run models
}
