% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_run.R
\name{luna_run}
\alias{luna_run}
\title{Run a NONMEM model}
\usage{
luna_run(id, folder = ".", type = "modelfit", ...)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{folder}{path to folder containing the model file. Default is current directory.}
}
\description{
Run a NONMEM model
}
