% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_run.R
\name{luna_run}
\alias{luna_run}
\title{Run a NONMEM model}
\usage{
luna_run(id, folder, method = NULL, diagnostic = NULL, verbose = TRUE, ...)
}
\arguments{
\item{id}{run id, e.g. \code{run1}. This will be the folder in which the NONMEM
model is run.}

\item{folder}{path to folder containing the model file. Default is current directory.}

\item{diagnostic}{run a diagnostic for the run, like a bootstrap or vpc. The
id \code{diagnostic} needs to be defined in the project yaml file.}
}
\description{
Run a NONMEM model
}
