% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna_tool.R
\name{luna_tool}
\alias{luna_tool}
\title{Run external tools (e.g. diagnostics) on models or run outputs
This can be used e.g. for bootstraps and VPCs. The function is implemented
in a modular way so that it can be easily extended.}
\usage{
luna_tool(
  id,
  tool,
  method = c("psn", "pharmpy"),
  force = FALSE,
  console = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{id}{run id}

\item{tool}{id for the tool, needs to be referenced in project YAML. See examples
for further details.}
}
\description{
Run external tools (e.g. diagnostics) on models or run outputs
This can be used e.g. for bootstraps and VPCs. The function is implemented
in a modular way so that it can be easily extended.
}
