% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tables_from_model.R
\name{remove_tables_from_model}
\alias{remove_tables_from_model}
\title{Remove all $TABLE records from a model}
\usage{
remove_tables_from_model(model, file)
}
\arguments{
\item{model}{pharmpy model object}

\item{variables}{character vector with variable names}

\item{firstonly}{add \code{FIRSTONLY} parameter to $TABLE record}
}
\description{
Remove all $TABLE records from a model
}
